refactor(codegen/golang): Pull opts into its own package
